Confirm Delete, add select and deselect all annotations

+++ src/client/src/components/side_panel/DeleteButton.tsx
@@ -1,6 +1,7 @@
 import { h } from "preact"
+import { useState } from "preact/hooks"
 import { ACTIONS } from "../../state/actions"
-import { get_all_selected_annotations } from "../../state/annotations/getters"
+import { get_selected_annotation_ids_owned_by_user } from "../../state/annotations/getters"
 
 import { State } from "../../state/state"
 import { get_store } from "../../state/store"
@@ -21,33 +22,57 @@ const connector = connect(map_state)
 
 function _DeleteButton (props: Props)
 {
-    const disabled = !props.selected_annotation_ids_owned_by_user
-    const fill = disabled ? "#aaa" : "#000"
+    const [confirm_delete, set_confirm_delete] = useState(false)
 
-    return <button
-        disabled={disabled}
-        title="Delete annotations"
-        onClick={() => {
-            const compound_ids = props.selected_annotation_ids_owned_by_user.split(",")
-            get_store().dispatch(ACTIONS.annotations.delete_annotations({ compound_ids }))
-        }}
-    >
-        <svg focusable="false" aria-hidden="true" viewBox="0 0 24 24" style={{ width: 20, fill }}>
-            <path d="M6 19c0 1.1.9 2 2 2h8c1.1 0 2-.9 2-2V7H6v12zM19 4h-3.5l-1-1h-5l-1 1H5v2h14V4z"></path>
-        </svg>
-    </button>
+    const disabled = !props.selected_annotation_ids_owned_by_user
+    const fill = disabled ? "#aaa" : (confirm_delete ? "#A00" : "#000")
+
+
+    return <div>
+        <button
+            disabled={disabled}
+            title="Delete annotations"
+            onClick={() =>
+            {
+                if (!confirm_delete)
+                {
+                    set_confirm_delete(true)
+                }
+                else
+                {
+                    const compound_ids = props.selected_annotation_ids_owned_by_user.split(",")
+                    get_store().dispatch(ACTIONS.annotations.delete_annotations({ compound_ids }))
+                    set_confirm_delete(false)
+                }
+            }}
+        >
+            <svg focusable="false" aria-hidden="true" viewBox="0 0 24 24" style={{ width: 20, fill }}>
+                <path d="M6 19c0 1.1.9 2 2 2h8c1.1 0 2-.9 2-2V7H6v12zM19 4h-3.5l-1-1h-5l-1 1H5v2h14V4z"></path>
+            </svg>
+
+            <br />
+            {confirm_delete && <span style={{ fontSize: 10, color: "#A00" }}>CONFIRM</span>}
+        </button>
+
+        {confirm_delete && <p>
+            <button
+                disabled={disabled}
+                title="Cancel delete"
+                onClick={() =>
+                {
+                    set_confirm_delete(false)
+                }}
+                style={{ width: "100%" }}
+            >
+                <svg focusable="false" aria-hidden="true" viewBox="0 0 24 24" style={{ width: 20, fill: "#000" }}>
+                    <path d="M6 19c0 1.1.9 2 2 2h8c1.1 0 2-.9 2-2V7H6v12zm2.46-7.12 1.41-1.41L12 12.59l2.12-2.12 1.41 1.41L13.41 14l2.12 2.12-1.41 1.41L12 15.41l-2.12 2.12-1.41-1.41L10.59 14l-2.13-2.12zM15.5 4l-1-1h-5l-1 1H5v2h14V4z"></path>
+                </svg>
+
+                <br />
+                <span style={{ fontSize: 10 }}>CANCEL</span>
+            </button>
+        </p>}
+    </div>
 }
 
 export const DeleteButton = connector(_DeleteButton)
-
-
-
-function get_selected_annotation_ids_owned_by_user(state: State)
-{
-    const { safe_user_name } = state.user
-    const ids = get_all_selected_annotations(state)
-        .filter(a => a && a.safe_user_name === safe_user_name)
-        .map(a => a!.compound_id)
-
-    return ids.join(",")
-}
